Move addr_prefixes.go
